refactor: 알림 API 응답 모델 변경

+++ android/app/src/main/java/com/mulkkam/data/remote/service/NotificationsService.kt
@@ -18,7 +18,7 @@ interface NotificationsService {
 
     @GET("/notifications")
     suspend fun getNotifications(
-        @Query("lastId") lastId: Int? = null,
+        @Query("lastId") lastId: Long? = null,
         @Query("clientTime") clientTime: String,
         @Query("size") size: Int,
     ): Result<NotificationsResponse>
